requirements.txt
Flask==2.3.2
PyJWT==2.7.0
requests==2.31.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
pywda==2.0.0
